AutomaticEmails.py v1.07
Version 1 is Done! Some final notes. To allow my script to access the email, I had to generate an app password for my computer. (It was a pain in the ass there was an annoying lack of documentation on the matter).  Otherwise the script works well. Some things I want to fix in Version 2, I want to make the graphic better. Adjusting to use RGBA values instead of RGB, as well as making the words more readable. Lastly, I was unable to use the API for games from the 2022-23 Season. It knew what games were being played but It wasn't able to find the scores or the records for said games. More to Come